--- app/assets/javascripts/decidim/comments/comments.js.erb
@@ -1,10 +1,0 @@
-// = require_self
-// = require ./bundle
-
-window.DecidimComments = window.DecidimComments || {};
-
-window.DecidimComments = {
-  assets: {
-    'icons.svg': "<%= asset_path "decidim/icons.svg" %>"
-  }
-};
